requirements.txt
click==7.1.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
python-dotenv==0.19.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
opencv_python==4.2.0.34
textblob==0.15.3
numpy==1.18.5
pytesseract==0.3.8
Pillow==8.3.2
